topology_generator/requirements.txt
# Core Dependencies
pydantic>=2.0.0
pyyaml>=6.0.0
lxml>=4.9.0

# CLI Support
argparse>=1.4.0

# Development and Testing
pytest>=7.3.0
pytest-cov>=4.0.0
mypy>=1.3.0
black>=23.3.0
flake8>=6.0.0

# Optional: Type checking
types-pyyaml>=6.0.0
types-setuptools>=67.8.0
